Comments seperate component added & seemore fix
