+++ src/components/Comment.jsx
@@ -1,0 +1,33 @@
+import { useState } from 'react';
+
+export function Comments({ data }) {
+  const [loadmore, setloadmore] = useState(true);
+  return (
+    <>
+      <div className="d-flex container bg-light mt-1 align-items-center">
+        <div className="">
+          {data.picture_url ? (
+            <div className="">
+              <img
+                src={data.picture_url}
+                className=" mt-2 small-round-pic me-2  round-img "
+                alt="..."
+              />
+            </div>
+          ) : (
+            <i className="mt-2 me-2 fa-solid fa-user fa-2x card-img-top small-round-pic  round-img bg-warning d-flex justify-content-center align-items-center"></i>
+          )}
+        </div>
+        <div className={loadmore && 'text-comment'}>{data.message}</div>
+      </div>
+      <div
+        className="seemore"
+        onClick={() => {
+          setloadmore((p) => !p);
+        }}
+      >
+        {loadmore ? 'Show More' : 'Show less'}
+      </div>
+    </>
+  );
+}
